Drop a- prefix for new pinning system
