--- packages/ceramic-core/src/store/a-state-store.ts
@@ -1,11 +1,0 @@
-import { DocState, Doctype } from "@ceramicnetwork/ceramic-common"
-
-export interface AStateStore {
-    open(): Promise<void>;
-    close(): Promise<void>;
-    save(document: Doctype): Promise<void>;
-    load(docId: string): Promise<DocState | null>;
-    exists(docId: string): Promise<boolean>;
-    list(docId?: string): Promise<string[]>;
-    remove(docId: string): Promise<void>;
-}
